requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
cloudant==2.14.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==2.3.1
django-environ==0.4.5
djangorestframework==3.12.2
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==5.4.0
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
identify==1.5.12
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.8
parameterized==0.8.1
Pillow==8.1.0
pluggy==0.13.1
pre-commit==2.9.3
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
shortuuid==1.0.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.3.0
zipp==3.4.0
